requirements.txt
fastapi==0.70.0
uvicorn==0.15.0

PyJWT==2.7.0
python-decouple==3.8